chore: designtech/ rename folder and add gitignore
