--- designtech/src/App.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import tech from './pages/tech.js';
-
-const App = () => {
-	  return <tech />;
-};
-
-export default App;
