+++ frontEnd/public/manifest.json
