+++ frontEnd/src/pages/Algorithm.js
