+++ frontEnd/src/pages/Architecture.js
